requirements.txt
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic[email]>=2.0.0

# Backend (FastAPI + Auth)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.10.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.20

# LangGraph + LangChain
langgraph
langchain
langchain-core
langchain-openai

# RAG — Vector Store
chromadb>=0.5.0

# Guardrails
nemoguardrails

# UI / Utilities
streamlit

# ML
sentence-transformers